That's quite the commit you got there

+++ src/main/java/com/dfsek/terra/biome/postprocessing/CoordinatePerturb.java
@@ -9,7 +9,7 @@ import org.polydev.gaea.math.FastNoiseLite;
 public class CoordinatePerturb {
     private final FastNoiseLite perturbX;
     private final FastNoiseLite perturbZ;
-    private final int amplitude;
+    private final double amplitude;
 
     /**
      * Create a CoordinatePerturb object with a given frequency, amplitude, and seed.
@@ -18,7 +18,7 @@ public class CoordinatePerturb {
      * @param amplitude Offset amplitude
      * @param seed      Noise seed
      */
-    public CoordinatePerturb(double frequency, int amplitude, long seed) {
+    public CoordinatePerturb(double frequency, double amplitude, long seed) {
         perturbX = new FastNoiseLite((int) seed);
         perturbX.setNoiseType(FastNoiseLite.NoiseType.OpenSimplex2);
         perturbX.setFrequency(frequency);
